Create common middleware
Merge entity and range middleware to common functionality.
