+++ src/services/mongo-express-middleware/middleware.js
@@ -1,18 +1,16 @@
 import { Router } from 'express';
-import mongoMethod from '../mongoMethod';
+import mongoMethod from './mongoMethod';
 import defaultCallback from './defaultCallback';
 
-const mapList = { create:'post', read:'get' };
-
-export default (model, ...rest) => {
+export default (model, map, ...rest) => {
   const router = Router({
     mergeParams: true,
   });
-
-  Object.keys(mapList).forEach(key => {
+  
+  Object.keys(map).forEach(key => {
     const callback = typeof rest[key] === 'function' ? rest[key] : defaultCallback;
     
-    router[mapList[key]]('/', (req, res, next) => {
+    router[map[key]]('/', (req, res, next) => {
       try {
         mongoMethod[key]({ req, res, next, callback, model, key });
       } catch(err) {
